backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.11.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.3
toml==0.10.2
tornado==6.2
whitenoise==6.3.0
